requirements.txt
azure-core==1.29.7
azure-cosmos==4.3.1
azure-functions==1.12.0
azure-storage-blob==12.14.1
stripe==5.5.0
azure-eventgrid==4.10.0
